+++ src/xrGame/game_sv_mp.cpp
@@ -1090,8 +1090,8 @@ void game_sv_mp::OnVoteStart(LPCSTR VoteCommand, ClientID sender)
 			string256 LevelName;
 			string256 LevelVersion;
 			sscanf_s(CommandParams, "%255s %255s",
-			         LevelName, sizeof(LevelName),
-			         LevelVersion, sizeof(LevelVersion)
+			         LevelName, (unsigned int)sizeof(LevelName),
+			         LevelVersion, (unsigned int)sizeof(LevelVersion)
 			);
 #ifdef DEBUG
 			Msg("--- Starting vote for changing level to: %s[%s]", LevelName, LevelVersion);
